Move ZoneFilterMock next to ZoneRegistryMock
